refactor: rename models file and fix path
