+++ backend/controllers/user-controller.js
@@ -1,4 +1,4 @@
-const User = require("../models/User");
+const User = require("../models/user");
 const createError = require("http-errors");
 
 const findAll = async (req, res) => {
